ELU (Exponential linear units) activation function added (http://arxiv.org/pdf/1511.07289v3.pdf). He et. al. weight initialization added (http://arxiv.org/pdf/1502.01852v1.pdf). for_ was raising a compile error on VC 2012 with TBB (due to type ambiguity). This is fixed via a cast operation.

+++ tiny_cnn/layers/fully_connected_layer.h
@@ -88,7 +88,7 @@ public:
             prev_delta[c] *= prev_h.df(prev_out[c]);
         }
 
-        for_(parallelize_, 0, out_size_, [&](const blocked_range& r) {
+        for_(parallelize_, 0, size_t(out_size_), [&](const blocked_range& r) {
             // accumulate weight-step using delta
             // dW[c * out_size + i] += current_delta[i] * prev_out[c]
             for (layer_size_t c = 0; c < in_size_; c++)
